backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
psutil
pywebview
gTTS
pygame
